# openapi.yaml
openapi: 3.0.0
info:
  title: Kallipolis ZK Security API
  version: 3.0.0
  description: |
    AI-Powered Security Layer for Polygon AggLayer.
    Features:
    - Real-time MEV detection
    - Smart contract audit (Slither, Mythril, Gemini)
    - Bridge verification
    - ZK proof generation
    - Formal verification
servers:
  - url: https://api.kallipolis.ai/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /firewall/check:
    post:
      summary: MEV firewall check
      description: Validates transaction for MEV attacks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transaction:
                  type: object
                  properties:
                    hash:
                      type: string
                    from:
                      type: string
                    to:
                      type: string
                    data:
                      type: string
                    gas:
                      type: integer
                    gasPrice:
                      type: integer
      responses:
        '200':
          description: Firewall verdict
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocked:
                    type: boolean
                  reasons:
                    type: array
                    items:
                      type: string
  /bridge/verify:
    post:
      summary: Bridge Merkle root verification
      description: Validates Merkle root for LxLy bridges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merkle_root:
                  type: string
                chain_id:
                  type: integer
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  checks:
                    type: object
                    properties:
                      format:
                        type: boolean
                      rpc:
                        type: object
                      zk_proof:
                        type: boolean
  /zk/prove:
    post:
      summary: Generate ZK proof
      description: Generates zero-knowledge proof for a circuit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                circuit:
                  type: string
                  enum: [mempool_firewall, bridge_merkle, balance_invariant]
                inputs:
                  type: object
      responses:
        '200':
          description: ZK proof
          content:
            application/json:
              schema:
                type: object
                properties:
                  proof:
                    type: string
                  public_signals:
                    type: array
                    items:
                      type: string
...